mental-health-backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pyodbc
pymysql
cryptography
passlib[bcrypt] 
python-jose[cryptography] 
fastapi-login